requirements.txt

openai==0.27.7
python-dotenv==1.0.0
streamlit==1.22.0
streamlit-option-menu==0.3.4
urllib3==1.26.6
